Moved commands in .cpp file and replace 'tilling' by 'tiling'
